Allows params with array values in button_to method
Updating the button_to helper method in ActionView::Helpers::UrlHelper to allow params option to handle array values.  Now, individual hidden input tags will be created for each element in the array. This replaces the former functionality of one hidden input element for the entire array with the values concatenated together in a slash-separated string.